requirement.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.6
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
